added ability to load custom model, by drag&drop
